py/pixel/requirements.txt
glob2
psutil 
tensorflow
keras
Pillow
torch
torchvision
torchaudio
matplotlib
pymongo[srv]

